refactor: Rename Poems to PoemShell for clarity of component use
