+++ src/components/App/App.js
@@ -1,7 +1,7 @@
 import React, {useState, useEffect} from 'react';
 import './App.css';
 import Form from '../Form/Form';
-import Poems from '../Poems/Poems';
+import PoemShell from '../Poems/PoemShell';
 import Home from '../Home/Home';
 import NotFound from '../NotFound/NotFound';
 import {getAllPoets, getPoems} from '../../utilities/apicalls';
@@ -57,7 +57,7 @@ const App = () => {
             <Route exact path='/select-poet'><Form poets={poets} setPoet={setPoet}/></Route>
             <Route exact path='/404'><NotFound /></Route>
             <Route exact path='/poem/:index' render={({match}) => {
-             return <Poems poems={poems} index={parseInt(match.params.index)}/>
+             return <PoemShell poems={poems} index={parseInt(match.params.index)}/>
             }}></Route>
             <Route path='*'><Redirect to='/404'/></Route>
           </Switch>
